Update Init files
We changed the shaders for beams so the init files need to be loaded
differently. Also, to make things easier for updating, the init_coop.lua
is gone since its the same file as init_NomadsCoop.lua

the difficulty is that now you need to rename the file to get it working
on the py client. ah well, im sure the users are able to figure this out
by following the instructions.