Mission 3 Improvements
Map auto expands from part 3 to part 4 using the Intel fleet frigate that spawns on the map and scouts the data centre for the playeers.

Part 4 now starts with a choice of either killing or capturing the data centres.
If you choose to kill them, you will get extra option to pay mass to nuke it.
If you choose to capture it, you will get free intel over the bases.